requirements.txt
apex==0.9.10dev
bitsandbytes==0.43.3
horovod==0.28.1
huggingface_hub==0.23.0
matplotlib==3.7.1
mmcv==1.6.1
numpy==1.24.3
onnxruntime==1.19.0
opencv_python==4.7.0.72
packaging==24.1
Pillow==10.4.0
PyYAML==6.0.2
safetensors==0.3.1
scikit_learn==1.2.2
tensorflow==2.17.0
tensorflow_datasets==4.9.6
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.12.2
wandb==0.17.7
webdataset==0.2.96
